Ajout de tests sscanf (0 / 6 passent...). #227

+++ applications/utils/include/unittest.h
@@ -1,7 +1,7 @@
 #ifndef __UNITTEST_H_
 #define __UNITTEST_H_
 
-void unit_test_int(const char* msg, int attendu, int obtenu);
+void unit_test_int(const char* msg, long long attendu, long long obtenu);
 void unit_test_str(const char* msg, const char* attendu, const char* obtenu);
 
 #endif
